Move content block components into subfolder
