+++ pages/_type/_slug.vue
@@ -55,7 +55,7 @@
 import Vue from 'vue'
 import { MetaInfo } from 'vue-meta'
 import { formatISO } from 'date-fns'
-import PostContentBlock from '../../components/PostContentBlock.vue'
+import PostContentBlock from '../../components/content/PostContentBlock.vue'
 import NewsletterSignup from '../../components/NewsletterSignup.vue'
 import {
   resolveAuthorLink,
